Version1.1/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
sqlalchemy==1.4.46
psycopg2-binary==2.9.6
pydantic== 2.0.1
pydantic-settings==2.0.3
python-dotenv==1.0.0
langchain==0.0.208
openai==0.27.8
transformers==4.29.2
torch==2.2.0
nltk==3.8.1
scikit-learn==1.2.2
httpx==0.24.1
langchain-openai==0.3.15
pytest==7.3.1
pytest-cov==4.1.0
pytest-asyncio==0.21.0
